Move tests to another dir
